requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gevent==20.9.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==0.4.17
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.1.2
